feat(clerkjs): Handle enterprise_sso response for sign in
Handle the case of a sign in request that includes an identifier that
matches the domain of an enterprise connection (SAML, Custom OAuth or
EASIE).

From version 2024-10-01 and onwards the API will respond with a status
of `needs_first_factor`, the only strategy that will be included in
supported first factors is `enterprise_sso` and the email address that
matched will be returned in the identifier field.

This commit updates the sign in component to read this new response and proceed
with the enterprise_sso strategy.